Added raw iLab rpm config client listing and initial rpm sorting.
The config client specifies what RPMs to install on client machines
outside of the configured RPM groups. I sorted them out based on what I
think is reasonable to replicate the current setup of the iLab machines.

Ideally we will generate the dockerfile based on the packages listed so
we can make adjustments as necessary.